Update migrations setup
Following the merging of the initial migration setup PR
(https://github.com/ministryofjustice/offender-management-allocation-api/pull/76)
it was established that there was already an entrypoint set in the
deployment.yaml file, and therefore made sense at this point to add the
db:migration command there.  As we get further into the workflow this
could well need to change, but suits our needs at present.