Adds support for generation of swagger docs and ui
This PR introduces a new dependency 'rswag' which will generate swagger
documentation from the rspec tests.  See spec/requests/health_spec.rb
for an example of usage, and https://github.com/domaindrivendev/rswag
for documentation.

Apart from the setup of rswag the main change here is that the
spec_helper now uses the rswag formatter as well as the default progress
formatter, which will generate the swagger file as the tests are run.

After running the rspec tests you can view the swagger ui at /api-docs