Add pre-commit hook to warn of errors git-crypt knows about
This gives us a bit of protection against not using git-crypt correctly.

The pre-commit hook needs to be installed as documented.

This could get slow because it's looking at all files in the repo rather than
only ones for which changes are being committed. I tried using a different
implementation which does look at only staged files but it wasn't working for
me:

    https://gist.github.com/Falkor/848c82daa63710b6c132bb42029b30ef