Allow tests to be auth'd
This PR allows rspec tests of the API requests to be auth'd and so we
can test the 200 responses we expect.

Tests can obtain a valid token for testing by adding the following to
the `response` section of the test.

```
let(:Authorization) { "Bearer #{generate_jwt_token}" }
```

To ensure that it fails with a bad token you can just set

```
let(:Authorization) { '' }
```

If you do not set the Authorization value at all, then the tests will
fail with an error.

As a side-effect of this feature we are also able to authenticate with a
token in the swagger-ui.