Adds a new /allocation endpoint for creating allocations
Creates a new endpoint at /allocation that can receive data and create a
new allocation between a POM and an offender.

The request should be a POST request, and contain the following in the
JSON object:

```
{
  "staff_id": "",
  "offender_no": "",
  "offender_id": "",
  "allocated_at_tier": "",
  "created_by": "",
  "prison": "",
  "reason": "",
  "note": "",
  "email":""
}
```

Currently the reason, note and email fields are optional.

The fields sent have specific meanings:

```
staff_id - The ID of the staff member who is having an offender
 	   allocated.

offender_no - The NOMIS ID of the offender

offender_id - The ID of the offender (which is different from the
	      offender_no).

allocated_at_tier - The offender's CURRENT tier.

created_by - The username of the person creating this allocation

prison - The prison code where the allocation is taking place

reason - The reason why the allocation has happened if it happened
	 against the recommendation provided.

note - A note on the allocation

email - The email address to receive a notification of allocation
```

If the response is a 200, then we have implemented a simple response
format (which may change) containing a status and errorMessage field to
determine whether the request was successful.  It is expected this may
acquire a `data` field once we are returning data.