Change Staff model to POMs
Following a discussion it was decided that we would be better served
renaming our Staff model to Prison Offender Managers.  We already fetch
staff details when prison staff are logging in, and so being explicit
about which particular staff we are referring to should help remove any
ambiguity and reduce confusion.  This commit is the start of this
process and covers the creation of the migration for changing the table
name as well as changing the model name.