Rename StaffService
This commit renames StaffService to PrisonOffenderManagerService to
match the new POM usage, and updates the AllocationService to ensure it
references this too.

There is also a migration to update the staff_id to nomis_staff_id as
this too had caused confusion as to which staff_id was being referred
to.