Add endpoint for getting offender's active allocations
We need to be able to determine whether an offender (or a list of
offenders) is allocated or not. This endpoint returns the active
allocation record for the provided offender_id (or list of offender_ids).

The response is the same as usual ..

```
{
  "status": "ok",
  "errorMessage": "",
  "data": {

  }
}

the data element contains a hash with the offender_id mapping to the
currently active allocation record (if any) for that offender. If the
offender is not currently allocated, their offender_id will not appear
as a key in the hash.