Update services to allow allocations to work
Whilst implementing the allocation work flow from within the allocation
manager we found a couple of issues that needed to be fixed for
allocations to work.  Firstly we needed to change the working_pattern to
be an integer rather than a string.  Secondly, we needed to pass in the
nomis_staff_id so had to ensure this was added to the params.